Test features

+++ Application/CartItem/DeleteCartItem/DeleteCartItemCommandHandler.cs
@@ -15,7 +15,10 @@ namespace EcommerceShop.Application.CartItem.DeleteCartItem
         {
             var item = await _repository.GetById(request.Id!);
 
-            await _repository.Delete(item);
+            if(item != null)
+            {
+                await _repository.Delete(item);
+            }
 
             return Unit.Value;
         }
